Ingest deadline testing database
