+++ tests/integration/hosts/maya/test_deadline_publish_in_maya.py
@@ -1,4 +1,3 @@
-from tests.lib.assert_classes import DBAssert
 from tests.integration.hosts.maya.lib import MayaDeadlinePublishTestClass
 
 
@@ -32,69 +31,6 @@ class TestDeadlinePublishInMaya(MayaDeadlinePublishTestClass):
     APP_VARIANT = ""
 
     TIMEOUT = 180  # publish timeout
-
-    def test_db_asserts(self, dbcon, publish_finished):
-        """Host and input data dependent expected results in DB."""
-        print("test_db_asserts")
-        failures = []
-        failures.append(DBAssert.count_of_types(dbcon, "version", 3))
-
-        failures.append(
-            DBAssert.count_of_types(dbcon, "version", 0, name={"$ne": 1}))
-
-        failures.append(
-            DBAssert.count_of_types(dbcon, "subset", 1,
-                                    name="modelMain"))
-
-        failures.append(
-            DBAssert.count_of_types(dbcon, "subset", 1,
-                                    name="renderTest_taskMain_beauty"))
-
-        failures.append(
-            DBAssert.count_of_types(dbcon, "subset", 1,
-                                    name="workfileTest_task"))
-
-        failures.append(DBAssert.count_of_types(dbcon, "representation", 8))
-
-        # hero included
-        additional_args = {"context.subset": "modelMain",
-                           "context.ext": "abc"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 2,
-                                    additional_args=additional_args))
-
-        # hero included
-        additional_args = {"context.subset": "modelMain",
-                           "context.ext": "ma"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 2,
-                                    additional_args=additional_args))
-
-        additional_args = {"context.subset": "modelMain",
-                           "context.ext": "mb"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 0,
-                                    additional_args=additional_args))
-
-        additional_args = {"context.subset": "renderTest_taskMain_beauty",
-                           "context.ext": "exr"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 1,
-                                    additional_args=additional_args))
-
-        additional_args = {"context.subset": "renderTest_taskMain_beauty",
-                           "context.ext": "jpg"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 1,
-                                    additional_args=additional_args))
-
-        additional_args = {"context.subset": "renderTest_taskMain_beauty",
-                           "context.ext": "png"}
-        failures.append(
-            DBAssert.count_of_types(dbcon, "representation", 1,
-                                    additional_args=additional_args))
-
-        assert not any(failures)
 
 
 if __name__ == "__main__":
